requirements.txt
Django==4.1.5
psycopg2==2.9.5
numpy==1.22.4
pandas==1.5.3
kaggle==1.5.12
djangorestframework==3.14.0
django-cors-headers==3.14.0